Added new Weapon Class and also added Colliders to Animations
Added colliders to some of the animations so that colliders are enabled during the last few frames of an animation, which means the player only takes damage in the end of the animation. This makes sense because precise and fair collision detection is essential in a souls-like game. Also refactored how collision works at the moment to coincide with this