Added more generic script for attack animation, and attacking in direction of mouse
Changed the way animation works for the player. Have a generic script which reads the clip name, and animates based on that. Player movement is disabled while attacking too. This is intentional mechanics.

Also changed how the player attacks. They no longer have to manually face a direction to attack, the animation simply plays based on the player's mouse position. e.g., If the mouse position is at the top of the screen, the player character plays an upward attack.