Fixed Colliders in Boss Room + Boss Attack created
Collider in the Boss Room example is now fixed to fit with node size so it doesn't caush game freezes. Also added a new basic attack for the boss which is to spawn Harmful objects in any position within the boss room thats above, below, to the left or to the right of the boss.