Made pathfinding more modular, Boss drops items and working on Inventory UI
Slightly changed pathfinding to work with any enemy rather than just 1 enemy which was used for testing purposes. Added a drop table script which chooses a random item for the enemy to drop. Will be changed to have rare drops as well like weapons. Finally started an outline of an inventory UI.