Added more animation and behaviour for the Ghost enemy
The ghost now has an animation where he throws his hands in the hair, which will correspond to him spawning small minions to attack the player later. The Ghost also flees if the player is too close, but it is slower than the player so that the player can catch and defeat it.