Added more to stat UI. Can now pause game
Added more Quality of Life updates to the stat UI. The upgrade panel now shows the player's current EXP and how much the upgrade costs. If the player can afford the upgrade, the text is green, and if they can't it's red. Player can now pause the game by pressing Space on Keyboard or Start on controller.