Added weapon equipping, cleaned up scripts
Adding a way to equip a weapon which changes the stats of the player's weapon. This is done by first collecting a weapon, and then clicking it in the inventory. Also cleaned up the scripts and made Item class abstract. ConsumableItem and WeaponItem now use this as a base class.