Added Lighting, More Inventory Functionality, New Enemy
Added Lighting to all the rooms and re-vamped the spawns to allow for new items. Added 4 New Weapons that can drop and be found in the levels. Also updated some icons and fonts throughout the game to be more consistent. Added a new Skeleton enemy.(WIP). Player can now drop items if they right click the item in their inventory.