Update UI, Added boss music, volume slider, saving volume
Overhauled the look of the UI. Added some music that plays whenever the boss starts attacking the player to add more atmosphere. Added a volume slider in the pause menu so that the player can change the volume levels. This is also saved and loaded if the player loads again.