Added Perk System with Scriptable objects, Some perks added
Added a new Perk class that will be extended to add more perks to weapons. Weapons now have an array of perks that can be easily added. Currently added a perk to modify health/stamina/movement speed and also a perk that instantiates an ally to attack the closest enemy relative to where it spawns.