Fixed some controller issues, Completed Note System
There was an issue using the Trigger button to attack where the player character would attack indefinitely when holding down the trigger which made for faster attacks as opposed to only attacking once when a button is pressed. Reverted attack button on controller to "X". Also completed the Note system that reads notes. Player can read previous notes as well and flick between ones they've seen. This will be updated to hold a counter of how many notes the player has seen so they don't see storyline from later on that they may not have enountered yet.